Generate default names for agents and networks on each save
Previously, the `default` values for the `name` fields of agents and networks were set using
a function with a prefix, assigned directly to the model field. However, because this function
was evaluated during migration generation, the same default value was reused for all rows
created after the same migration. This also caused Django to detect spurious changes in
migrations, as it expected the default value to be fixed.

What we actually want is to generate a new name for each individual row when it's saved.
To achieve this, we now move the default value logic into the model's `save()` method, so
a fresh name is generated at save time only if the field is unset.

Signed-off-by: dodo920306 <dodo920306@gmail.com>