Make agent URLs as CharField validated by a customized validator
As an application which largely rely on modern microsvc deployment architecture as
provided by tools like docker and k8s, Cello must be able to be compatible with all of
them, which means to be compatible with lots of components within an internal
container network. Thus, the ability to accept internal URL references, like container
and service names, compling with RFC-1123
(https://www.rfc-editor.org/rfc/rfc1123.html) is indeed necessary.

This commit introduces a new dependency FQDN (https://github.com/ypcrts/fqdn) in
order to achieve this very goal. A hostname which is accepted by this tool should be
seen as valid by Cello. Moreover, in order to introduce this customized validator, the
urls field of `Agent` model is now changed into CharField to stop serializers of Django
REST framework to use the default method to validate the field. An overriding in the
serializers level can be done but must be done for all `Agent` serializers respectively.
Therefore, to keep code simplicity, this commit directly changes the field type.
Changing the validator of the field all at once. As a result, the URL of Agent should be
fully compatible with an internal container network after this commit.

Signed-off-by: dodo920306 <dodo920306@gmail.com>